Proof-of-concept for removing the logger from `RequestHandler<T>`.
`RequestHandler` took a logger into its constructor,
so that when it continues the pipeline by calling `_successor.Handle`
it's able to log the names of the current handler and the next handler.

We found this painful, because when we converted to Brighter every
single one of our handlers had to take an `ILog` dependency - and it
was _Brighter's_ `ILog`, not the logging interface of our choice.
(Even if we were using LibLog, we still wouldn't be able to use the copy
of `LibLog` that we'd imported.) So the options were:
  * Have _two_ loggers (of different interfaces) injected into every
    handler - one so that Brighter can log pipeline continuations and
    another so that we can do our application-level logging.
  * Abandon our old logging interface and use the _Brighter_ `ILog`
    everywhere.

Neither option is particularly palatable. In my opinion, the logger
should be injected into Brighter _once_ at the entry point to the
whole system (when we build the command processor), not at
every handler.

So with that in mind, I've excised the `ILog` from `RequestHandler`
altogether. I replaced it with a more generalised solution - an `event`
(a C# `event`, not a Brighter `Event`) on `IHandleRequests` to indicate
that the pipeline is continuing. `RequestHandler` invokes that event when
it calls `_successor.Handle`, and propagates calls from its successor up
the pipeline. In `CommandProcessor` itself, the event is bound to the
actual logger.