Replace `Successor` write-only property with `SetSuccessor` method.
We had problems with our IoC container - it recognised that `Successor`
was a settable property with a type of `IHandleRequests<T>` and helpfully
tried to inject that property. This resulted in a loop of successors.

We worked around it eventually, but here's a patch replacing that
public setter with a method, in the hope that IoC containers won't
try and do something dumb with it.