Fix #132 by serialising access to the Producer.
I made the return type of `MessagingGateway.Send` into `void`
(instead of `Task`). The `Task` was always immediately being
awaited.

This is going to break backwards compatibility with anyone that was
using or implementing `IAmAMessageProducer`, but @iancooper doesn't
think there is anyone doing that, so it's a risk we are willing to take.

The "proper" way to do it is to restrict access to the
RabbitMQ connection to a single thread (an actor).
`Send` would ask the actor to publish the message and return a
`Task` which will be completed when the actor has successfully
published it. We decided it wasn't worth the effort of going down
this route yet, since there's no `PostAsync` on the `CommandProcessor`
in the first place; instead we are simply using a critical section.
That's something you'll have to deal with in the future.