Use update expression to mark messages as dispatched
This changes `MarkDispatchedAsync` to do so using an Update expression. This enables the ability to do it using a single call to the DB, instead of reading everything out and writing it all back in again.

To facilitate this, a reference to the Dynamo DB client has been added as a member variable, as update expressions can only be used when using the low level API. This has also meant deleting one of the (unused) constructors that takes the Dynamo DB context directly, as the client is not accessible from within the context object.

Finally, the ADR has been updated to reflect this change as this wasn't the original proposed change.