Fix observability tests
This fixes the observability tests, as they were looking for events in the wrong place (they're now on the per-operation create span instead of the per-message clear span as we pre-fetch all required messages in a batch). It also adds the missing span name for the new Count operation on outboxes.